+++ mrb_doc/textures.rb
@@ -24,12 +24,3 @@ def draw_texture_pro(texture, source, destination, origin, rotation, colour)
   # src/mruby_integration/textures.cpp
   nil
 end
-
-# Returns a new colour which is a faded version of the original.
-# @param colour [Colour]
-# @return [Colour]
-def fade(colour)
-  # mrb_fade
-  # src/mruby_integration/textures.cpp
-  Colour.new
-end