+++ src/mruby_integration/textures.cpp
@@ -52,30 +52,6 @@ mrb_draw_texture_pro(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_nil_value();
 }
 
-auto
-mrb_fade(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Color* colour;
-  mrb_float alpha;
-  mrb_get_args(mrb, "df", &colour, &Colour_type, &alpha);
-
-  auto* return_colour = static_cast<Color*>(malloc(sizeof(Color)));
-  *return_colour = Fade(*colour, alpha);
-
-  mrb_value obj = mrb_obj_value(
-    Data_Wrap_Struct(mrb, Colour_class, &Colour_type, return_colour));
-
-  setup_Colour(mrb,
-               obj,
-               return_colour,
-               return_colour->r,
-               return_colour->g,
-               return_colour->b,
-               return_colour->a);
-
-  return obj;
-}
-
 void
 append_textures(mrb_state* mrb)
 {
@@ -86,6 +62,4 @@ append_textures(mrb_state* mrb)
                     "draw_texture_pro",
                     mrb_draw_texture_pro,
                     MRB_ARGS_REQ(6));
-
-  mrb_define_method(mrb, mrb->kernel_module, "fade", mrb_fade, MRB_ARGS_REQ(2));
 }