+++ CHANGELOG.md
@@ -260,6 +260,8 @@
 - Texture2D#generate_texture_mipmaps is now C++ backed
 - Removed generate_texture_mipmaps
 - Reworked some rake tasks into shell scripts
+- Added Colour#fade, Colour#fade!, and Colour#inspect
+- Removed fade
 
 ## v0.3.14.1